+++ apps/prairielearn/src/tests/helperServer.ts
@@ -5,6 +5,7 @@ import { setTimeout as sleep } from 'node:timers/promises';
 import { assert } from 'chai';
 import * as opentelemetry from '@prairielearn/opentelemetry';
 import debugfn from 'debug';
+import { cache } from '@prairielearn/cache';
 
 import * as assets from '../lib/assets';
 import { config } from '../lib/config';
@@ -13,7 +14,6 @@ import * as cron from '../cron';
 import * as socketServer from '../lib/socket-server';
 import * as serverJobs from '../lib/server-jobs-legacy';
 import * as freeformServer from '../question-servers/freeform';
-import * as cache from '../lib/cache';
 import * as localCache from '../lib/local-cache';
 import * as codeCaller from '../lib/code-caller';
 import * as externalGrader from '../lib/externalGrader';
@@ -74,7 +74,11 @@ export function before(courseDir: string = TEST_COURSE_PATH): () => Promise<void
       socketServer.init(httpServer);
 
       debug('before(): initialize cache');
-      await cache.init();
+      await cache.init({
+        type: config.cacheType,
+        keyPrefix: config.cacheKeyPrefix,
+        redisUrl: config.redisUrl,
+      });
 
       debug('before(): initialize server jobs');
       serverJobs.init();
@@ -103,14 +107,14 @@ export async function after(): Promise<void> {
     debug('after(): stop server');
     await promisify(server.stopServer)();
 
+    debug('after(): close socket server');
+    await socketServer.close();
+
     debug('after(): close load estimators');
     load.close();
 
     debug('after(): stop cron');
     await cron.stop();
-
-    debug('after(): close socket server');
-    await socketServer.close();
 
     debug('after(): close server jobs');
     await serverJobs.stop();