+++ packages/aws/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @prairielearn/aws
 
+## 1.0.7
+
+### Patch Changes
+
+- dce0fa3: Upgrade all JavaScript dependencies
+
 ## 1.0.6
 
 ### Patch Changes