+++ packages/signed-token/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @prairielearn/signed-token
 
+## 1.0.12
+
+### Patch Changes
+
+- dce0fa3: Upgrade all JavaScript dependencies
+
 ## 1.0.11
 
 ### Patch Changes